result updates

+++ src/main/java/Repository/AccountRepositoryIMPL.java
@@ -3,10 +3,7 @@ package Repository;
 import ConnectWithBD.ConnectWithBD;
 import Model.AcountingRecords;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.Statement;
+import java.sql.*;
 import java.util.Date;
 
 public class AccountRepositoryIMPL implements AccountRepository {
@@ -29,7 +26,11 @@ public class AccountRepositoryIMPL implements AccountRepository {
             preparedStmt.setDate(5, (java.sql.Date) returnDate);
             preparedStmt.setString(6, status);
             preparedStmt.execute();
-        } catch (Exception e) {
+        }
+        catch (SQLIntegrityConstraintViolationException ex) {
+            System.out.println("Аккаунт с таким ID уже существует");
+        }
+        catch (Exception e) {
             System.out.println(e);
         }
     }
@@ -37,9 +38,19 @@ public class AccountRepositoryIMPL implements AccountRepository {
     @Override
     public void deleteAccount(Integer accountId) {
         try (Connection con = ConnectWithBD.getConnection()) {
-            PreparedStatement pstmt = con.prepareStatement("DELETE FROM accountingrecords where accountId = ?");
-            pstmt.setInt(1, accountId);
-            pstmt.executeUpdate();
+            stmt = con.createStatement();
+            String query = ("select * FROM accountingrecords where accountId = ?");
+            PreparedStatement pst = con.prepareStatement(query);
+            pst.setInt(1, accountId);
+            rs = pst.executeQuery();
+            if(rs.next()) {
+                PreparedStatement pstmt = con.prepareStatement("DELETE FROM accountingrecords where accountId = ?");
+                pstmt.setInt(1, accountId);
+                pstmt.executeUpdate();
+            }
+            else{
+                System.out.println("Аккаунта с таким ID не существует");
+            }
         }
         catch (Exception e) {
             e.printStackTrace();
@@ -50,14 +61,23 @@ public class AccountRepositoryIMPL implements AccountRepository {
     public void modifyAccount(Integer accountId, Integer updateClientId, Integer updateBookId, Date updateReceiptDate, Date updateReturnDate, String UpdateStatus) {
         try (Connection con = ConnectWithBD.getConnection()) {
             stmt = con.createStatement();
-            String query = "update accountingrecords set clientId = ?, bookId = ?, returnDate = ?, status = ?  where accountId = ?";
-            PreparedStatement preparedStmt = con.prepareStatement(query);
-            preparedStmt.setInt(1, updateClientId);
-            preparedStmt.setInt(2, updateBookId);
-            preparedStmt.setDate(3, (java.sql.Date) updateReturnDate);
-            preparedStmt.setString(4, UpdateStatus);
-            preparedStmt.setInt(5, accountId);
-            preparedStmt.executeUpdate();
+            String query1 = ("select * FROM accountingrecords where accountId = ?");
+            PreparedStatement pst = con.prepareStatement(query1);
+            pst.setInt(1, accountId);
+            rs = pst.executeQuery();
+            if(rs.next()) {
+                String query = "update accountingrecords set clientId = ?, bookId = ?, returnDate = ?, status = ?  where accountId = ?";
+                PreparedStatement preparedStmt = con.prepareStatement(query);
+                preparedStmt.setInt(1, updateClientId);
+                preparedStmt.setInt(2, updateBookId);
+                preparedStmt.setDate(3, (java.sql.Date) updateReturnDate);
+                preparedStmt.setString(4, UpdateStatus);
+                preparedStmt.setInt(5, accountId);
+                preparedStmt.executeUpdate();
+            }
+            else{
+                System.out.println("Аккаунта с таким ID не существует");
+            }
         } catch (Exception e) {
             System.out.println(e);
         }
@@ -71,16 +91,22 @@ public class AccountRepositoryIMPL implements AccountRepository {
             PreparedStatement preparedStmt = con.prepareStatement(query);
             preparedStmt.setInt(1, accountId);
             rs = preparedStmt.executeQuery();
-            while (rs.next()) {
-                int id = rs.getInt(1);
-                int clientId = rs.getInt(2);
-                int BookId = rs.getInt(3);
-                Date date1 = rs.getDate(4);
-                Date date3 = rs.getDate(5);
-                String status = rs.getString(6);
-                selectAccount = new AcountingRecords(id,clientId,BookId,date1, date3,status);
+            boolean z = rs.next();
+            if(z) {
+                while (z) {
+                    int id = rs.getInt(1);
+                    int clientId = rs.getInt(2);
+                    int bookId = rs.getInt(3);
+                    Date date1 = rs.getDate(4);
+                    Date date3 = rs.getDate(5);
+                    String status = rs.getString(6);
+                    selectAccount = new AcountingRecords.AccountBuilder(id).setClientId(clientId).setBookId(bookId).setReceiptDate(date1).setReturnDate(date3).setStatusId(status).build();
+                    z = false;
+                }
             }
-
+            else {
+                System.out.println("Аккаунта с таким ID не существует");
+            }
         } catch (Exception e) {
             System.out.println(e);
         }
