+++ src/main/java/Repository/AuthorsRepositoryIMPL.java
@@ -3,10 +3,7 @@ package Repository;
 import ConnectWithBD.ConnectWithBD;
 import Model.Authors;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.Statement;
+import java.sql.*;
 
 public class AuthorsRepositoryIMPL  implements  AuthorsRepository{
 
@@ -24,7 +21,10 @@ public class AuthorsRepositoryIMPL  implements  AuthorsRepository{
             preparedStmt.setString(2, newAuthorFirstname);
             preparedStmt.setString(3, newAuthorLastname);
             preparedStmt.execute();
-        } catch (Exception e) {
+        }
+        catch (SQLIntegrityConstraintViolationException ex){
+            System.out.println("Автор с таким ID уже существует");
+        }catch (Exception e) {
             System.out.println(e);
         }
     }
@@ -32,9 +32,19 @@ public class AuthorsRepositoryIMPL  implements  AuthorsRepository{
     @Override
     public void deleteAuthor(Integer authorId) {
         try (Connection con = ConnectWithBD.getConnection()) {
-            PreparedStatement pstmt = con.prepareStatement("DELETE FROM authors where authorId = ?");
-            pstmt.setInt(1, authorId);
-            pstmt.executeUpdate();
+            stmt = con.createStatement();
+            String query = ("select * FROM authors where authorId = ?");
+            PreparedStatement pst = con.prepareStatement(query);
+            pst.setInt(1, authorId);
+            rs = pst.executeQuery();
+            if(rs.next()) {
+                PreparedStatement pstmt = con.prepareStatement("DELETE FROM authors where authorId = ?");
+                pstmt.setInt(1, authorId);
+                pstmt.executeUpdate();
+            }
+            else{
+                System.out.println("Автора с таким ID не существует");
+            }
         }
         catch (Exception e) {
             e.printStackTrace();
@@ -45,12 +55,21 @@ public class AuthorsRepositoryIMPL  implements  AuthorsRepository{
     public void modifyAuthor(Integer authorId, String updateAuthorFirstname, String updateAuthorLastname) {
         try (Connection con = ConnectWithBD.getConnection()) {
             stmt = con.createStatement();
-            String query = "update authors set firstname = ?, lastname = ? where authorId = ?";
-            PreparedStatement preparedStmt = con.prepareStatement(query);
-            preparedStmt.setString(1, updateAuthorFirstname);
-            preparedStmt.setString(2, updateAuthorLastname);
-            preparedStmt.setInt(3, authorId);
-            preparedStmt.executeUpdate();
+            String query1 = ("select * FROM authors where authorId = ?");
+            PreparedStatement pst = con.prepareStatement(query1);
+            pst.setInt(1, authorId);
+            rs = pst.executeQuery();
+            if(rs.next()) {
+                String query = "update authors set firstname = ?, lastname = ? where authorId = ?";
+                PreparedStatement preparedStmt = con.prepareStatement(query);
+                preparedStmt.setString(1, updateAuthorFirstname);
+                preparedStmt.setString(2, updateAuthorLastname);
+                preparedStmt.setInt(3, authorId);
+                preparedStmt.executeUpdate();
+            }
+            else{
+                System.out.println("Автора с таким ID не существует");
+            }
         } catch (Exception e) {
             System.out.println(e);
         }
@@ -63,11 +82,18 @@ public class AuthorsRepositoryIMPL  implements  AuthorsRepository{
             PreparedStatement preparedStmt = con.prepareStatement(query);
             preparedStmt.setInt(1, authorId);
             rs = preparedStmt.executeQuery();
-            while (rs.next()) {
-                int id = rs.getInt(1);
-                String firstname = rs.getString(2);
-                String lastname = rs.getString(3);
-                selectAuthor = new Authors(id,firstname,lastname);
+            boolean z = rs.next();
+            if(z) {
+                while (z) {
+                    int id = rs.getInt(1);
+                    String firstname = rs.getString(2);
+                    String lastname = rs.getString(3);
+                    selectAuthor = new Authors.AuthorsBuilder(id).setFirstName(firstname).setLastName(lastname).build();
+                    z = false;
+                }
+            }
+            else {
+                System.out.println("Автора с таким ID не существует");
             }
 
         } catch (Exception e) {
