+++ src/main/java/Repository/BooksRepositoryIMPL.java
@@ -3,10 +3,7 @@ package Repository;
 import ConnectWithBD.ConnectWithBD;
 import Model.Books;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.Statement;
+import java.sql.*;
 
 public class BooksRepositoryIMPL implements BooksRepository {
 
@@ -26,7 +23,11 @@ public class BooksRepositoryIMPL implements BooksRepository {
             preparedStmt.setInt(3, newAuthorId);
             preparedStmt.setString(4, newCategory);
             preparedStmt.execute();
-        } catch (Exception e) {
+        }
+        catch (SQLIntegrityConstraintViolationException ex){
+            System.out.println("Книга с таким ID уже существует");
+        }
+        catch (Exception e) {
             System.out.println(e);
         }
     }
@@ -35,10 +36,21 @@ public class BooksRepositoryIMPL implements BooksRepository {
     public void deleteBook(Integer bookId) {
 
         try (Connection con = ConnectWithBD.getConnection()) {
-            PreparedStatement pstmt = con.prepareStatement("DELETE FROM books where idBook = ?");
-            pstmt.setInt(1, bookId);
-            pstmt.executeUpdate();
+            stmt = con.createStatement();
+            String query = ("select * FROM books where idBook = ?");
+            PreparedStatement pst = con.prepareStatement(query);
+            pst.setInt(1, bookId);
+            rs = pst.executeQuery();
+            if(rs.next()) {
+                PreparedStatement pstmt = con.prepareStatement("DELETE FROM books where idBook = ?");
+                pstmt.setInt(1, bookId);
+                pstmt.executeUpdate();
+            }
+            else{
+                System.out.println("Книги с таким ID не существует");
+            }
         }
+
         catch (Exception e) {
             e.printStackTrace();
         }
@@ -49,17 +61,28 @@ public class BooksRepositoryIMPL implements BooksRepository {
 
         try (Connection con = ConnectWithBD.getConnection()) {
             stmt = con.createStatement();
-            String query = "update clients set bookName = ?, authorId = ?, categoryOfBook = ? where idBook = ?";
-            PreparedStatement preparedStmt = con.prepareStatement(query);
-            preparedStmt.setString(1, updateBookName);
-            preparedStmt.setInt(2, updateAuthorId);
-            preparedStmt.setString(3, updateCategory);
-            preparedStmt.setInt(4, bookId);
-            preparedStmt.executeUpdate();
+            String query1 = ("select * FROM books where idBook = ?");
+            PreparedStatement pst = con.prepareStatement(query1);
+            pst.setInt(1, bookId);
+            rs = pst.executeQuery();
+            if(rs.next()) {
+                String query = "update books set bookName = ?, authorId = ?, categoryOfBook = ? where idBook = ?";
+                PreparedStatement preparedStmt = con.prepareStatement(query);
+                preparedStmt.setString(1, updateBookName);
+                preparedStmt.setInt(2, updateAuthorId);
+                preparedStmt.setString(3, updateCategory);
+                preparedStmt.setInt(4, bookId);
+                preparedStmt.executeUpdate();
+            }
+            else{
+                System.out.println("Книги с таким ID не существует");
+            }
+
         } catch (Exception e) {
             System.out.println(e);
         }
     }
+
     @Override
     public Books searchTheBooks(int booksId) {
 
@@ -69,14 +92,20 @@ public class BooksRepositoryIMPL implements BooksRepository {
             PreparedStatement preparedStmt = con.prepareStatement(query);
             preparedStmt.setInt(1, booksId);
             rs = preparedStmt.executeQuery();
-            while (rs.next()) {
-                int id = rs.getInt(1);
-                String firstname = rs.getString(2);
-                int idAuthor = rs.getInt(3);
-                String lastname = rs.getString(4);
-                selectBook = new Books(id,firstname,idAuthor,lastname);
+            boolean z = rs.next();
+            if(z) {
+                while (z) {
+                    int id = rs.getInt(1);
+                    String firstname = rs.getString(2);
+                    int idAuthor = rs.getInt(3);
+                    String category = rs.getString(4);
+                    selectBook = new Books.BooksBuilder(id).setName(firstname).setAuthorId(idAuthor).setCategory(category).build();
+                    z = false;
+                }
             }
-
+            else {
+                System.out.println("Книги с таким ID не существует");
+            }
         } catch (Exception e) {
             System.out.println(e);
         }
