+++ src/main/java/Repository/ClientRepositoryIMPL.java
@@ -2,15 +2,8 @@ package Repository;
 
 import ConnectWithBD.ConnectWithBD;
 import Model.Client;
-import Model.SpisokOfEntity;
-
-import java.awt.print.Book;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.Statement;
-import java.util.Iterator;
-import java.util.Map;
+import java.io.IOException;
+import java.sql.*;
 
 public class ClientRepositoryIMPL implements ClientRepository {
 
@@ -28,17 +21,33 @@ public class ClientRepositoryIMPL implements ClientRepository {
             preparedStmt.setString(2, newFirstname);
             preparedStmt.setString(3, newLastname);
             preparedStmt.execute();
-        } catch (Exception e) {
+        }
+        catch (SQLIntegrityConstraintViolationException ex){
+            System.out.println("Клиент с таким ID уже существует");
+        }
+        catch (SQLException | IOException e) {
             System.out.println(e);
         }
+
     }
 
     public void deleteClient(Integer clientId) {
 
         try (Connection con = ConnectWithBD.getConnection()) {
-            PreparedStatement pstmt = con.prepareStatement("DELETE FROM clients where idClients = ?");
-            pstmt.setInt(1, clientId);
-            pstmt.executeUpdate();
+            stmt = con.createStatement();
+            String query = ("select * FROM clients where idClients = ?");
+            PreparedStatement pst = con.prepareStatement(query);
+            pst.setInt(1, clientId);
+            rs = pst.executeQuery();
+               if(rs.next()){
+                   PreparedStatement pstmt = con.prepareStatement("DELETE FROM clients where idClients = ?");
+                   pstmt.setInt(1, clientId);
+                   pstmt.executeUpdate();
+               }
+               else{
+                   System.out.println("Клиента с таким ID не существует");
+               }
+
         }
         catch (Exception e) {
             e.printStackTrace();
@@ -50,12 +59,22 @@ public class ClientRepositoryIMPL implements ClientRepository {
 
        try (Connection con = ConnectWithBD.getConnection()) {
            stmt = con.createStatement();
-           String query = "update clients set firstname = ?, lastname = ? where idClients = ?";
-           PreparedStatement preparedStmt = con.prepareStatement(query);
-           preparedStmt.setString(1, updateFirstname);
-           preparedStmt.setString(2, updateLastname);
-           preparedStmt.setInt(3, clientId);
-           preparedStmt.executeUpdate();
+           String query1 = ("select * FROM clients where idClients = ?");
+           PreparedStatement pst = con.prepareStatement(query1);
+           pst.setInt(1, clientId);
+           rs = pst.executeQuery();
+           if(rs.next()) {
+               String query = "update clients set firstname = ?, lastname = ? where idClients = ?";
+               PreparedStatement preparedStmt = con.prepareStatement(query);
+               preparedStmt.setString(1, updateFirstname);
+               preparedStmt.setString(2, updateLastname);
+               preparedStmt.setInt(3, clientId);
+               preparedStmt.executeUpdate();
+           }
+           else{
+                   System.out.println("Клиента с таким ID не существует");
+               }
+
        } catch (Exception e) {
            System.out.println(e);
        }
@@ -69,14 +88,21 @@ public class ClientRepositoryIMPL implements ClientRepository {
             PreparedStatement preparedStmt = con.prepareStatement(query);
             preparedStmt.setInt(1, clientId);
             rs = preparedStmt.executeQuery();
-            while (rs.next()) {
-                int id = rs.getInt(1);
-                String firstname = rs.getString(2);
-                String lastname = rs.getString(3);
-                selectClient = new Client(id,firstname,lastname);
-            }
-
-        } catch (Exception e) {
+            boolean z = rs.next();
+               if(z){
+                while (z) {
+                    int id = rs.getInt(1);
+                    String firstname = rs.getString(2);
+                    String lastname = rs.getString(3);
+                    selectClient = new Client.ClientBuilder(id).setFirstName(firstname).setLastName(lastname).build();
+                    z = false;
+                }
+}
+               else {
+                   System.out.println("Клиента с таким ID не существует");
+               }
+        }
+        catch (Exception e) {
             System.out.println(e);
         }
         return selectClient;
